This is the first version of the program, though it's pretty far along.
1) I mostly used the "read data" functions from BaselineFunctions.py, but I made minor changes and added "Quarterly" to their names.

I created a process for reading each quarter's data separately, but aggregating all of the quarters into a single set of files stored in a subdirectory called QtrAggregate.

I created the episode triggering logic though it still needs to include a check for the 59 day rule on Part D claims.

I've added logic to find the end date of episodes based on the 6 month rule, hospice start date (which should be separate for MEOS and PBP) and death date.

I created the MEOS audit program and have used and tested in on CCSIL.