many changes made to nearly finish the reports.
Among the changes, I fixed the filterdfepi function to use dates.  That broke the filterdfepi for the baseline period because the dates don't overlap, obviously.  I removed filterdfepi for the reports that include baseline data.  That was a mistake.  I needed, instead, to create filterdfepiB which filters on physician and cancer type, but not on dates.  I added that function but have only applied it to the end of life tab so far.